Split up scene_loader
