+++ src/ElgForward/material.h
@@ -8,7 +8,10 @@ struct Material {
   PixelShaderHandle PixelShader;
 };
 
-inline bool CreateMaterial(const std::string& id, const filesystem::path& vs_path, const filesystem::path& ps_path, ID3D11Device* device, Material* material) {
+inline bool CreateMaterial(const std::string& /* id */, const std::string& /* type */, const filesystem::path& /* base_path */, ID3D11Device* /* device */, Material* /* material */) {
+  // basic -> vs.cso, ps.cso
+
+  /*
   material->Hash = std::hash<std::string>()(id);
 
   material->VertexShader = CreateVertexShader(vs_path, std::unordered_map<std::string, VertexDataChannel>(), device);
@@ -20,6 +23,7 @@ inline bool CreateMaterial(const std::string& id, const filesystem::path& vs_pat
   if (!material->PixelShader.IsValid()) {
     return false;
   }
+  */
 
   return true;
 }
