+++ src/ElgForward/mesh.cpp
@@ -8,6 +8,7 @@
 
 #pragma warning(push)
 #pragma warning(disable: 4201)
+#pragma warning(disable: 4305)
 #include <assimp/cimport.h>
 #include <assimp/scene.h>
 #include <assimp/postprocess.h>
